Code refactor, add predicted labels saving, add training wrapper
- Added predicted labels saving
- Added train wrapper
- Refactored code and put some impls into .cpp file
- Removed old basic examples
- Changed labels type to unsigned int everywhere

+++ src/statistics/accuracy.h
@@ -18,7 +18,7 @@ public:
      * @param expected - expected labels
      * @return accuracy of NN predictions
      */
-    auto static accuracy(const std::vector<size_t> &predicted, const std::vector<size_t> &expected){
+    auto static accuracy(const std::vector<unsigned int> &predicted, const std::vector<unsigned int > &expected){
         float correctPredictions = 0;
         for (size_t i = 0; i < predicted.size(); i++){
             if (predicted[i] == expected[i]){
