+++ src/statistics/argmax.h
@@ -18,7 +18,7 @@ public:
      * @return - vector of classes
      */
     auto static argmax(const Matrix<float> &matrix){
-        auto classes = std::vector<size_t>(matrix.getNumRows());
+        auto classes = std::vector<unsigned int>(matrix.getNumRows());
         for (size_t i = 0; i < matrix.getNumRows(); i++){
             float currentMax = 0;
             for (size_t j = 0; j < matrix.getNumCols(); j++){
